Refactor users CRUD to SQLAlchemy
