+++ src/backend/db/__init__.py
@@ -1,2 +1,3 @@
 from flask_sqlalchemy import SQLAlchemy
-db = SQLAlchemy
+
+db = SQLAlchemy()
